Move toolbar functionality into App.vue
It is planned for the functions in the context menu to also be quickly accessible in the toolbar. Currently, the toolbar functionality is split into multiple files, which makes it difficult to implement and maintain, since each .vue file needs to provide functionality for both toolbar and context menu. This commit aims to solve this by including such functionality within App.vue, enabling common functionality between the context menu and the toolbar.