Add functionality to context menu in App
Implement basic context menu functionality by adding to the `<template>`, `data()` and `methods()` of `App`.

The `ContextMenu` component alone will not work - some way of using the shared functionality between the context menu and the toolbar like `addShape()` needs to exist. This commit only adds to `App.vue` (and eventually this newly added code will need to be moved to another component) but the `<template>` contains the intended structure of the context menu.