Improve closing of context menu
Before this commit, the context menu was only closed by right-clicking (or more specifically, invoking the `oncontextmenu` event). Now, the context menu can be closed by simply clicking anything outside of the context menu.

The opening and closing of the context menu was also handled by `openContextMenu()`; this commit moves the handling of the closing of the context menu in another method `closeContextMenu()`.