Add ability to tweak numerical parameters in EditArea
By selecting a number in the editing area and holding Shift, the selected number can then be changed by moving the cursor. The numbers that can be selected are the strings which can be converted to a `Number`.

Releasing Shift stops the changing of the selected number, and updates the shape's style property.