Fix currID resetting on every refresh of the page
The code implementing the fix takes the shapeList from localStorage and sets currID as the maximum ID out of the existing shapes plus one. If there are no shapes in the localStorage shapeList then currID is set to 1 instead.

The catch block in the try-catch within the start of data() has also been filled with a console.error statement, just in case of an error.