Refactor code in App
As suggested by the Vue docs and `eslint`, a `v-bind` for the `key` attribute should be present within a `v-for`. This is addressed in this commit.

The indentation for the region of the code corresponding to the `div#display` node has also been fixed in accordance with the code style of this program.