Add prototype of shape list
Shapes may become invisible, for example by setting `visibility: hidden` or `display: none`. So this commit introduces a list, which has an entry for each shape present, displaying the ID and name.

There is currently no interaction with the list other than that clicking on the name of a shape in this list opens the CSS edit window for that shape.