Add ShapeRender
As both the div used for shape rendering and editor content change upon any update of the corresponding style attribute, the changes in the varySelection() method of EditArea cannot be reflected straight away in the render of the shape. But better control of the rendering can be achieved by separating the shape renderer from App into its own component, hence this commit.