Implement update of shape render on varySelection
This commit updates the shape's appearance upon varying values within the corresponding EditArea through selecting a number, holding shift and then moving mouse.

The current implementation within this commit requires both the ShapeRender and the corresponding EditArea to be grouped together by some tag (in this case, a <div>). It updates the ShapeRender directly with the new text within the editable area.