Change id to class in EditArea
The value of the `id` attribute on a tag should be unique. This commit fixes the common value of the id attribute present in all instances of the `EditArea` component by turning id value `edit-region` into a class name instead.

The selectors for `edit-region` that are used within the methods of the `EditArea` component have also been changed accordingly.