Refactor to separate context menu to its own component
The template and methods of the context menu took a large amount of space in `App.vue`. Also, the specific code of the context menu was quickly becoming tighly coupled with the code of `App`. Hence the context menu has been moved to its own component called `ContextMenu` in this commit.

Changes to the opening and closing of the context menu had to be made. Instead of listening to particular mouse events in the `#app` element, event listeners for certain actions are created instead upon mount of `ContextMenu`.

Additionally, since the entries within the menu call on methods present only in the parent (such as `copyShape`, `editShape`), appropriate `$emit` and `v-on` bindings are introduced in this commit to call them.