Edit ContextMenu to appear directly under cursor
The context menu, before this commit, appeared a few pixels away from the cursor upon right clicking. This may be potentially confusing for users who are expecting the actual browser context menu after right clicking twice. This commit changes this so that the upper right corner of the context menu appears right below the cursor.