Change <body> to have overflow:hidden
This commit makes it so that the scrollbar does not appear whenever an element takes space outside of the `#display` element. This happened, for example, when a shape was so large it extended out of the display, or when the context menu was opened near the edge of the display.