Refactor second 'add shape' test in frontend.js
The second test for adding shapes checks whether this can be done through the context menu. This test has been refactoed in this commit to use the page objects introduced earlier.

This test makes use of a user action in nightwatch.js, namely `.mouseButtonClick()`, and such things are apparently absent from the page object instance. As a temporary workaround, `.mouseButtonClick()` is instead invoked from `browser` instead of the `visualtool` page object instance.