Refactor e2e test code to improve code style
There was some difficulties in the second 'add shape' test case in accessing `mouseButtonClick` from the page object instance, and this method was invoked from `browser` instead as a workaround. Upon further research, a reference to `browser` is available through the `api` property of the page object instance (see [https://nightwatchjs.org/api/pageobject/#page-object-instance]). This commit thus replaces the use of `browser` with `visualtool.api` in the "'add shape' via context menu" test.

Accordingly, the style of the test code had to be changed.