Remove @menuOpenCollection in Context Menu test
The `@menuOpenCollection` element was supposed to be introduced in a future commit, however, it has been accidentally included in the current version. This means that the Context Menu test suite does not currently work as intended. Hence, this commit brings a removal to the assertion for the visibility of `@menuOpenCollection` within `tests/context-menu.js` for the time being.