Add tests for the future `Collection` component
The `Collection` test suite added in this commit checks for the basic functionalities of the Collection component (to be added in a future commit). Appropriate elements in the visualtool page object has also been included.

The `dragAndDrop()` command in the `visualtool` page object has been changed in this commit to include an alternative usage: `dragAndDrop(sourceSelector, destinationSelector)` drag and drops from the position of the source selector, and into the destination selector. This is used within the Collection test suite.

A menu entry which reads 'open collection' is also now expected in the context menu.