Fix logic and exception handling in visualtool.js
The error in the logic of checking whether less than 2 or more than 3 arguments were passed in 'dragAndDrop()' has been fixed (by substitution of `&&` in place of `||`).

Also, the exception handling for an incorrect number of arguments passed to `dragAndDrop()` was displaying incorrectly in nightwatch. It was enough to change the thrown value from a `String` to an `Error` object, so this fix is included in this commit.