Add Collection component
This commit adds the Collection component into the application. It also includes several other necessary changes to integrate it into the application, and to ensure that all but the last assertion of the test "A shape can be dragged and dropped to the collection, saving it" (within tests/collection.js) passes.

To implement drag and drop of visible shapes into the collection, a new property called `shapeHeld` has been introduced, whose value is filled with the shape held upon dragging a visible ShapeRender.

Two context menu entries, one to open the collection, and another to close the collection, have also been added.