Add ContextMenuItem component
The ContextMenuItem component added in this commit tidies the <template> contents of ContextMenu.

To properly integrate this component into visualtool, the methods deleteShape, copyShape and editShape have been changed to take in a reference of the shape node itself. An `action()` method has also been added to ContextMenu, which currently wraps the `$emit()` and `closeContextMenu()` calls for use in the `@click` of each ContextMenuItem.