Improve conditions for 'open/close collection'
The context menu items 'open collection' and 'close collection' were previously visible on certain contexts, regardless of the visibility of the collection itself. This commit has the addition of a `collectionVisible` prop to ContextMenu so that the visibility of the collection is available to the ContextMenu. This means that the 'open/close collection' ContextMenuItems can conditionally appear depending on whether the collection is open or not.