Change v-if to v-show in ContextMenuItem
Previously, ContextMenuItems only rendered when the condition held true within the v-if. This made the "The collection cannot be opened twice with context menu" test fail for the wrong reason. This commit changes v-if to v-show to keep the rendering of all ContextMenuItem elements whilst toggling visibility based on the value of the condition, so that now the aforementioned test passes.