Add display export feature
This commit adds the ability to export the rendered shapes in the display onto a HTML file.

The user can click the "export" button on the toolbar and download the resulting "export.html". visualtool then copies the rendered shapes from the display to the HTML file.

Due to personal commitments, I currently only have time to do manual testing of the program, so this feature has been added without any testing in advance.