Remove dependence in methods in ContexMenu tests
This commit removes all but one instance of calls to methods in the ContextMenu tests so that the component interface is tested, instead of its implementation. The tests now expect the ContextMenu to manage visibility and context through its props, instead of being done internally as it currently is.

The call to the `action` method of ContextMenu may be removed later once an alternative is found. However, it remains for now as no other solution could be found.