Replace the visibility unit tests for ContextMenu
It was found that `document.dispatchEvent` could be used to simulate a right click on the `document` itself, which removes the need for using props to control the visibility of the `ContextMenu`. This removes the need for having props in ContextMenu, and also allows testing the current version of ContextMenu that changes visibility upon right click on the document. This current version also attempts to prevent coupling of the ContextMenu to the root App component.