Add initial code for EditArea unit tests
This commit adds some unit tests for the EditArea component. These test mainly check for the behaviour of the component in different settings when mounting, and specifies some of the behaviour when interacting with it.

One unit test still remains to be added, and refactoring of the test code to remove or enumerate the hard-coded indices needs to be done, since randomised testing with random `shape` objects may be desired in the future. For reference, an instance of hard-coded indices can be found in the "updates style upon ctrl+enter" test.