Add EditArea unit test for varying parameters
This commit adds to EditArea.test.js a unit test which checks whether the varying of parameters works correctly. This unit test is relatively large, although it is included for now since, in the sense that all necessary events are specified exactly in the test code, it is more well-defined than the e2e test for EditArea fulfilling the same purpose.