Add test for App's toolbar element
This commit adds a unit test to App's toolbar which checks whether the buttons with specified labels exist in the toolbar. This test suite currently lacks the unit tests for the correct actions being run by each button, which has been deferred to a future commit as a vuex migration is planned. Migration to vuex may also allow for testing for the actions that these buttons execute.