Add the makeShapeList utility function
In this commit, `makeShapeList()` is introduced as an export of `src/utils.js` to enable creation of `shapeList`s initialised with random shape objects. The `shapeList` variables which were previously explicitly initialised in `Display.test.js` and `ShapeList.test.js` have been changed to use the `makeShapeList()` function, now generating a `shapeList` of length 20 instead of 3.

This commit does not however contain the complete set of all possible testable CSS values for each property in the style of a randomly generated shape. This is because some values such as `rgba(0, 0, 0, 1)` are translated into `rgb(0, 0, 0)` by JSDOM after assigning the style string to a cssText attribute. As a result, such values have been ommitted for this commit.