Fix makeShapeList presets for EditArea unit test
The shapeList for the EditArea unit test "can modify an arbitrary number by selecting text, holding shift, then moving the mouse around" needs some CSS property in the shape to have the value '50px'. Hence this commit adds a preset for the style prop of the shape in this test.